Refactored style code so it works when minified
The previous implementation did not work when class names were minified.

Also in this refactoring, I merged meas.ts and meas-type.ts, which was
necessary to avoid circular dependencies, but I've wanted to do this for
a while anyway to bring them more in line with the other template+object
classes.