Dropped a To-Do
I *think* the original idea of this To-Do was that a Measurement that is
created by `MeasListEditor` and then opened in a `MeasurementEditor`
doesn't have the `isNew` flag set in that editor anymore, however, I now
think that users who are using the `MeasListEditor` are unlikely to need
the `MeasurementEditor`. It's also possible I was confusing the `isNew`
flag with the `skipInitWarns` parameter of `warningsCheck` when I wrote
this (or the older `isBrandNew` flag), but currently `isNew` is *only*
used for UI stuff: deciding where to scroll to and whether to expand
text boxes. So taking these two things together, I think this To-Do can
be removed, and only if I notice some undesired visual UI behavior I
could consider reinstating it.